+++ package/vite.config.ts
@@ -1,25 +1,26 @@
+// vite.config.ts
 import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
 import pluginRewriteAll from 'vite-plugin-rewrite-all';
+import tsconfigPaths from 'vite-tsconfig-paths';
 import svgr from 'vite-plugin-svgr';
 import dts from 'vite-plugin-dts';
 
 export default defineConfig(({ command }) => ({
   plugins: [
-    dts(),
+    dts({ insertTypesEntry: true }),
     react(),
-
-    /** Allows us importing svg's */
+    tsconfigPaths({
+      projects: ['./tsconfig.json']
+    }),
     svgr(),
-
-    /** Created an override on the server to redirect all requests to the dev server to index.html (so react-router can handle it) */
     pluginRewriteAll(),
   ],
   build: {
     emptyOutDir: false,
-    sourcemap: command === 'build' ? true : 'inline', // build for prod(command "build") we want source map in seperated files. during dev we ok with inline source map
+    sourcemap: command === 'build' ? true : 'inline',
     rollupOptions: {
-      external: ['React', 'react', 'react-dom'],
+      external: ['react', 'react-dom'],
       output: {
         // Provide global variables to use in the UMD build
         // for externalized deps
@@ -32,10 +33,7 @@ export default defineConfig(({ command }) => ({
       entry: './index.ts',
       name: '@transmitsecurity/riskid-reactjs-ts',
       formats: ['es', 'cjs'],
-      fileName: format => {
-        return `index.${format}.js`;
-      }
+      fileName: format => `index.${format}.js`,
     },
   }
-}),
-);
+}));